atoi strchr strnstr strrchr autop

+++ Libft/ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: mravera <@student.42lausanne.ch>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/23 13:24:18 by mravera           #+#    #+#             */
-/*   Updated: 2021/11/24 17:59:29 by mravera          ###   ########.fr       */
+/*   Updated: 2021/12/02 03:01:26 by mathis           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,18 +14,18 @@
 
 char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 {
-	int	i;
-	int	j;
+	size_t  i;
+	size_t  j;
 
 	j = 0;
 	i = 0;
 	if (*needle == 0)
 		return ((char *)haystack);
-	while (haystack[i] != 0 && i < (int)len)
+	while (haystack[i] != 0 && i < len)
 	{
 		while (haystack[i + j] == needle[j])
 		{
-			if (needle[j + 1] == 0 && (i + j < (int)len))
+			if (needle[j + 1] == 0 && (i + j < len))
 				return (&((char *)haystack)[i]);
 			else
 				j++;
